chore: removed deprecated fn refactor: moved util
